% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypothyroid}
\alias{hypothyroid}
\title{Hypothyroid}
\format{
An object of class \code{data.frame} with 2772 rows and 23 columns.
}
\usage{
hypothyroid
}
\description{
Hypothyroid
}
\keyword{datasets}
